% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{data.as.matrix}
\alias{data.as.matrix}
\title{Convert data.frame to matrix}
\usage{
data.as.matrix(
  x,
  formula = sampleName ~ regionName,
  which.col.score = "score",
  ...
)
}
\arguments{
\item{x}{}

\item{formula}{casting formula. See \code{\link{reshape2::acast}} for specifics.}

\item{which.col.score}{name of column which stores values for scoring. See `value.var` in \code{\link{reshape2::acast}} for specifics.}
}
\value{

}
\description{
Convert data.frame to matrix given formula and
}
\examples{

}
\author{

}
