% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-util.R
\name{dt.as_interval}
\alias{dt.as_interval}
\title{Convert data.table as `data.table.interval` object}
\usage{
dt.as_interval(
  x,
  by = NULL,
  which.col.start = "start",
  which.col.end = "end",
  extra.col.rm = FALSE
)
}
\arguments{
\item{x}{}

\item{by}{}

\item{extra.col.rm}{}
}
\value{

}
\description{
(TBA)
}
\details{
Copied from package Xmisc (beta)
}
\examples{

}
\author{

}
