% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-util.R
\name{dt.merge_interval}
\alias{dt.merge_interval}
\title{Merge data.table.interval}
\usage{
dt.merge_interval(x, y, by = NULL, ...)
}
\arguments{
\item{x}{}

\item{y}{}

\item{by}{character to grouped by, i.e. "refid", "chr"}
}
\value{
data.table
}
\description{
(TBA)
}
\examples{

}
\author{
Xiaobei Zhao
}
