% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-util.R
\name{dt.score_interval}
\alias{dt.score_interval}
\title{To score `data.region` using `data.score`}
\usage{
dt.score_interval(
  data.region,
  data.score,
  by = c("chr"),
  which.col.start = "start",
  which.col.end = "end",
  which.col.score = "score",
  which.col.sampleName = "sampleName",
  which.col.regionName = "regionName",
  FUNC.SCORE = NULL,
  score.weighted = FALSE,
  pcThreads = get_DEFAULT__pcThreads()
)
}
\arguments{
\item{data.region}{a data table of regions to score. i.e. a data table of region information by feature (e.g. gene).}

\item{data.score}{a data table of score information over regions. i.e. a data table of score information over regions by sample.}

\item{by}{the column, within each level of which `data.region` and `data.score` need be processed individually. (e.g. chromosome)}

\item{which.col.start}{the column for the star coordinate of a region for overlapping. (0-based)}

\item{which.col.end}{the column for the end coordinate of a region for overlapping. (1-based)}

\item{which.col.score}{the column for score}

\item{which.col.sampleName}{the column for sample identifiers}

\item{which.col.regionName}{the column for region (feature, e.g. gene) identifiers}

\item{FUNC.SCORE}{a function to compute a single score given a vector of scores. (e.g. mean or median)
see \code{\link{get_FUNC.SCORE_mean}}, \code{\link{get_FUNC.SCORE_median}}}

\item{pcThreads}{the processing threads. (pcThreads > 1 is not supported on Windows)}
}
\value{
a data table with region name, coordinates, sample name and scores over a region per sample.
}
\description{
To score `data.region` using `data.score` by subgrouped regions.
}
\details{
In computational biology, it scores a set of regions across chromosomes using a set of scores by overlapping genomic coordinates.
}
\author{
Xiaobei Zhao
}
