% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-util.R
\name{dt.sort_interval}
\alias{dt.sort_interval}
\title{Sort data.table.interval}
\usage{
dt.sort_interval(x, by = NULL, ...)
}
\arguments{
\item{x}{}

\item{by}{}
}
\value{
data.table
}
\description{
(TBA)
}
\author{
Xiaobei Zhao
}
