% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{get_FUNC.SCORE.WEIGHTED_mean}
\alias{get_FUNC.SCORE.WEIGHTED_mean}
\title{A wrapper function to compute weighted mean}
\usage{
get_FUNC.SCORE.WEIGHTED_mean(x, size, na.rm = TRUE)
}
\arguments{
\item{x}{}

\item{size}{}
}
\value{

}
\description{
A wrapper function to compute weighted mean, weighted by size
}
\author{

}
