% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{get_FUNC.SCORE_mean}
\alias{get_FUNC.SCORE_mean}
\title{A wrapper function of `mean`}
\usage{
get_FUNC.SCORE_mean(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
A wrapper function of `mean`, with default of removing NA values
}
\author{
get_FUNC.SCORE_mean(c(1:6,99))
}
