% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{get_FUNC.SCORE_median}
\alias{get_FUNC.SCORE_median}
\title{A wrapper function of `median`}
\usage{
get_FUNC.SCORE_median(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
A wrapper function of `median`, with default of removing NA values
}
\author{
get_FUNC.SCORE_median(c(1:6,99))
}
