% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{get_FUNC.SCORE_vconcat}
\alias{get_FUNC.SCORE_vconcat}
\title{A wrapper function of `vconcat`}
\usage{
get_FUNC.SCORE_vconcat(x)
}
\value{

}
\description{
(TBA)
}
\examples{
get_FUNC.SCORE_vconcat(c(1:6,99))

}
\author{

}
