% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{prep_data.region}
\alias{prep_data.region}
\title{To prepare `data.region`}
\usage{
prep_data.region(
  x,
  which.col.chr = "chr",
  which.col.start = "start",
  which.col.end = "end",
  which.col.regionName = "name"
)
}
\arguments{
\item{x}{}

\item{which.col.chr}{}

\item{which.col.start}{}

\item{which.col.end}{}

\item{which.col.name}{}
}
\value{

}
\description{
`data.region` is a data table of regions to score.
}
\details{
e.g. a data table of region information by feature (e.g. gene).
}
\author{

}
