% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{prep_data.score}
\alias{prep_data.score}
\title{To prepare `data.score`}
\usage{
prep_data.score(
  x,
  which.col.chr = "chr",
  which.col.start = "start",
  which.col.end = "end",
  which.col.sampleName = "name",
  which.col.score = "score"
)
}
\arguments{
\item{x}{}

\item{which.col.chr}{colname for chromosomes or other sequences}

\item{which.col.start}{colname for the start of a region}

\item{which.col.end}{colname for the end of a region}

\item{which.col.score}{colname for the score of a region}

\item{which.col.name}{colname for the name of a region (e.g. sample names)}
}
\value{

}
\description{
`data.score` is a data table of score information over regions.
}
\details{
e.g. a data table of score information over regions by sample.
}
\author{

}
