% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsdata-internal.R
\name{read.data}
\alias{read.data}
\title{Read data as a `data.table`}
\usage{
read.data(inFpath, header = TRUE, ...)
}
\arguments{
\item{inFpath}{file, See \code{read.table}}

\item{header}{See \code{read.table}}

\item{...}{}
}
\value{

}
\description{
Read a file in `data.table` format
}
\examples{

}
\author{

}
